Improvements in the way the text size is calculated
When moved the input in the card editor to the top, broken the text size
recalculation. The code was fragile, because identified the node by position.
Now we use the class name to get the elements where the text size need be recalculated.
Is similar to how was implemented at the beggining, but now we have multiple
classes defined, then a parameter is used to select the class.

The second change is not decresse the font size by one, but by 1/10 of the actual size.
This make faster the resize, and as a aditional advantage, the text is in general
a little smaller than the box, showing a little padding.